+++ include/kernel/application/syscalls.h
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,25 +29,47 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <jinue/shared/asm/errno.h>
-#include <kernel/domain/entities/descriptor.h>
-#include <kernel/domain/entities/object.h>
-#include <kernel/domain/entities/process.h>
-#include <kernel/domain/syscalls.h>
+#ifndef JINUE_KERNEL_APPLICATION_SYSCALLS_H
+#define JINUE_KERNEL_APPLICATION_SYSCALLS_H
 
-int close(int fd) {
-    descriptor_t *desc;
-    int status = dereference_object_descriptor(&desc, get_current_process(), fd);
+#include <kernel/types.h>
 
-    if(status < 0) {
-        return status;
-    }
+int close(int fd);
 
-    object_header_t *object = desc->object;
-    
-    close_object(object, desc);
+int create_endpoint(int fd);
 
-    desc->flags = DESCRIPTOR_FLAG_NONE;
+int create_process(int fd);
 
-    return 0;
-}
+int create_thread(int  process_fd, const thread_params_t *params);
+
+int destroy(int fd);
+
+int dup(int process_fd, int src, int dest);
+
+void exit_thread(void);
+
+void *get_thread_local(void);
+
+int get_user_memory(const jinue_buffer_t *buffer);
+
+int mclone(int src, int dest, const jinue_mclone_args_t *args);
+
+int mint(int owner, const jinue_mint_args_t *mint_args);
+
+int mmap(int process_fd, const jinue_mmap_args_t *args);
+
+int puts(int loglevel, const char *str, size_t length);
+
+void reboot(void);
+
+int receive(int fd, jinue_message_t *message);
+
+int reply(const jinue_message_t *message);
+
+int send(int fd, int function, const jinue_message_t *message);
+
+void set_thread_local(void *addr, size_t size);
+
+void yield_thread(void);
+
+#endif
