+++ kernel/application/kmain.c
