+++ kernel/application/syscalls/create_thread.c
@@ -1,22 +1,22 @@
 /*
- * Copyright (C) 2019-2024 Philippe Aubertin.
+ * Copyright (C) 2024 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,9 +29,29 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <jinue/shared/asm/errno.h>
+#include <kernel/application/syscalls.h>
+#include <kernel/domain/entities/descriptor.h>
+#include <kernel/domain/entities/object.h>
+#include <kernel/domain/entities/process.h>
 #include <kernel/domain/entities/thread.h>
-#include <kernel/domain/syscalls.h>
 
-void yield_thread(void) {
-    switch_thread_yield();
+int create_thread(int  process_fd, const thread_params_t *params) {
+    descriptor_t *desc;
+    int status = dereference_object_descriptor(&desc, get_current_process(), process_fd);
+
+    if(status < 0) {
+        return status;
+    }
+
+    process_t *process = get_process_from_descriptor(desc);
+
+    if(process == NULL) {
+        return -JINUE_EBADF;
+    }
+
+    const thread_t *thread = construct_thread(process, params);
+
+    /** TODO associate new thread to a free descriptor */
+    return (thread == NULL) ? -JINUE_EAGAIN : 0;
 }
