+++ kernel/application/syscalls/destroy.c
@@ -1,22 +1,22 @@
 /*
- * Copyright (C) 2019-2024 Philippe Aubertin.
+ * Copyright (C) 2024 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -30,25 +30,30 @@
  */
 
 #include <jinue/shared/asm/errno.h>
-#include <jinue/shared/asm/syscall.h>
-#include <kernel/domain/services/logging.h>
-#include <kernel/domain/syscalls.h>
+#include <kernel/application/syscalls.h>
+#include <kernel/domain/entities/descriptor.h>
+#include <kernel/domain/entities/object.h>
+#include <kernel/domain/entities/process.h>
 
-int puts(int loglevel, const char *str, size_t length) {
-    if(length > JINUE_PUTS_MAX_LENGTH) {
-        return -JINUE_EINVAL;
+int destroy(int fd) {
+    descriptor_t *desc;
+    int status = dereference_object_descriptor(&desc, get_current_process(), fd);
+
+    if(status < 0) {
+        return status;
     }
 
-    switch(loglevel) {
-    case JINUE_PUTS_LOGLEVEL_INFO:
-    case JINUE_PUTS_LOGLEVEL_WARNING:
-    case JINUE_PUTS_LOGLEVEL_ERROR:
-        break;
-    default:
-        return -JINUE_EINVAL;
+    if(!descriptor_is_owner(desc)) {
+        return -JINUE_EPERM;
     }
 
-    logging_add_message(loglevel, str, length);
+    object_header_t *object = desc->object;
+
+    close_object(object, desc);
+
+    mark_object_destroyed(object);
+
+    desc->flags = DESCRIPTOR_FLAG_NONE;
 
     return 0;
 }
