+++ kernel/application/syscalls/mclone.c
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,9 +29,64 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <kernel/domain/syscalls.h>
-#include <kernel/machine/memory.h>
+#include <jinue/shared/asm/errno.h>
+#include <kernel/application/syscalls.h>
+#include <kernel/domain/entities/descriptor.h>
+#include <kernel/domain/entities/process.h>
+#include <kernel/machine/vm.h>
 
-int get_user_memory(const jinue_buffer_t *buffer) {
-    return machine_get_memory_map(buffer);
+/**
+ * Implementation for the MCLONE system call
+ *
+ * Clone memory mappings from one process to another.
+ *
+ * @param src source process descriptor number
+ * @param dest destination process descriptor number
+ * @param args MCLONE system call arguments structure
+ * @return zero on success, negated error code on failure
+ *
+ */
+int mclone(int src, int dest, const jinue_mclone_args_t *args) {
+    process_t *current_process = get_current_process();
+
+    descriptor_t *src_desc;
+    int status = dereference_object_descriptor(&src_desc, current_process, src);
+    
+    if(status < 0) {
+        return status;
+    }
+    
+    process_t *src_process = get_process_from_descriptor(src_desc);
+
+    if(src_process == NULL) {
+        return -JINUE_EBADF;
+    }
+    
+    descriptor_t *dest_desc;
+    status = dereference_object_descriptor(&dest_desc, current_process, dest);
+    
+    if(status < 0) {
+        return status;
+    }
+    
+    process_t *dest_process = get_process_from_descriptor(dest_desc);
+
+    if(dest_process == NULL) {
+        return -JINUE_EBADF;
+    }
+
+    bool success = machine_clone_userspace_mapping(
+        dest_process,
+        args->dest_addr,
+        src_process,
+        args->src_addr,
+        args->length,
+        args->prot
+    );
+
+    if(!success) {
+        return -JINUE_ENOMEM;
+    }
+
+    return 0;
 }
