+++ kernel/application/syscalls/puts.c
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -30,31 +30,25 @@
  */
 
 #include <jinue/shared/asm/errno.h>
-#include <jinue/shared/asm/permissions.h>
-#include <kernel/domain/entities/descriptor.h>
-#include <kernel/domain/services/ipc.h>
-#include <kernel/domain/syscalls.h>
-#include <kernel/machine/thread.h>
+#include <jinue/shared/asm/syscall.h>
+#include <kernel/application/syscalls.h>
+#include <kernel/domain/services/logging.h>
 
-int receive(int fd, jinue_message_t *message) {
-    thread_t *receiver = get_current_thread();
-
-    descriptor_t *desc;
-    int status = dereference_object_descriptor(&desc, receiver->process, fd);
-
-    if(status < 0) {
-        return status;
+int puts(int loglevel, const char *str, size_t length) {
+    if(length > JINUE_PUTS_MAX_LENGTH) {
+        return -JINUE_EINVAL;
     }
 
-    ipc_endpoint_t *endpoint = get_endpoint_from_descriptor(desc);
-
-    if(endpoint == NULL) {
-        return -JINUE_EBADF;
+    switch(loglevel) {
+    case JINUE_PUTS_LOGLEVEL_INFO:
+    case JINUE_PUTS_LOGLEVEL_WARNING:
+    case JINUE_PUTS_LOGLEVEL_ERROR:
+        break;
+    default:
+        return -JINUE_EINVAL;
     }
 
-    if(!descriptor_has_permissions(desc, JINUE_PERM_RECEIVE)) {
-        return -JINUE_EPERM;
-    }
+    logging_add_message(loglevel, str, length);
 
-    return receive_message(endpoint, receiver, message);
+    return 0;
 }
