+++ kernel/application/syscalls/reboot.c
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,31 +29,10 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <jinue/shared/asm/errno.h>
-#include <kernel/domain/entities/descriptor.h>
-#include <kernel/domain/entities/object.h>
-#include <kernel/domain/entities/process.h>
-#include <kernel/domain/syscalls.h>
+#include <kernel/application/syscalls.h>
+#include <kernel/machine/halt.h>
 
-int destroy(int fd) {
-    descriptor_t *desc;
-    int status = dereference_object_descriptor(&desc, get_current_process(), fd);
-
-    if(status < 0) {
-        return status;
-    }
-
-    if(!descriptor_is_owner(desc)) {
-        return -JINUE_EPERM;
-    }
-
-    object_header_t *object = desc->object;
-
-    close_object(object, desc);
-
-    mark_object_destroyed(object);
-
-    desc->flags = DESCRIPTOR_FLAG_NONE;
-
-    return 0;
+void reboot(void) {
+    /* TODO check permissions */
+    machine_reboot();
 }
